refactor(ai-tools): extract getByResource, searchByKeyword, timeline to operation-handlers/ticket-specialty
